% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{biome_layer}
\alias{biome_layer}
\title{Categorical raster.}
\format{
A RasterLayer
\describe{
\item{class}{RasterLayer}
\item{dimensions}{4, 3, 12   (nrow, ncol, ncell)}
\item{resolution}{10, 10  (x, y)}
\item{extent}{11.71845, 40.85081, -47.89832, -4.428839  (xmin, xmax, ymin, ymax)}
\item{crs}{"+proj=longlat +datum=WGS84 +no_defs"}
\item{source}{memory}
\item{names}{layer}
\item{values}{Numerical categorical value.}
}
}
\description{
Categorical RasterLayer cropped to southern Africa at 10 degree resolution.
Represents a biome classification RasterLayer that could be used when spatially blocking species occurrence data.
}
