% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamic_proj.R
\name{dynamic_proj}
\alias{dynamic_proj}
\title{Project species distribution and abundance models onto dynamic environmental covariates.}
\usage{
dynamic_proj(
  dates,
  projection.method,
  local.directory = NULL,
  drive.folder = NULL,
  user.email = NULL,
  sdm.mod = NULL,
  sdm.thresh = 0.5,
  sdm.weight = 1,
  sam.mod = NULL,
  sam.weight = 1,
  save.directory = NULL,
  save.drive.folder = NULL
)
}
\arguments{
\item{dates}{a character string, vector of dates in format YYYY-MM-DD.}

\item{projection.method}{a character string or vector, the method or methods to project distribution and abundance onto projection covariates. Options include "'proportional'", "'binary'", "'abundance'" and "'stacked'". See details for more information.}

\item{local.directory}{optional; a character string, the path to a local directory to read projection covariate data frames from.}

\item{drive.folder}{optional; a character string, the Google Drive folder to read projection covariate data frames from.}

\item{user.email}{optional; a character string, user email for initialising Google Drive. Required if drive.folder or save.drive.folder used.}

\item{sdm.mod}{optional; a model object or list of model objects fitted to species distribution data.}

\item{sdm.thresh}{optional; a numeric value, the threshold to convert projected distribution suitability into binary presence-absence. Default 0.5. Required if projection.method is "'binary'" or "stacked".}

\item{sdm.weight}{optional; a numeric string, weights given to each sdm.mod model projection, given in the same order as the sdm.mod list. Default is equal weighting to all models.}

\item{sam.mod}{optional; a model object or list of model objects fitted to species abundance data.}

\item{sam.weight}{optional; a numeric string, weights given to each sdm.mod model projection, given in the same order as the sam.mod list. Default is equal weighting to all models.}

\item{save.directory}{optional; a character string, path to local directory to save projection rasters to.}

\item{save.drive.folder}{optional; a character string, Google Drive folder to save projection rasters to.}
}
\value{
Exports model projection rasters for each projection date to user-specified Google Drive folder or local directory.
}
\description{
Projects fitted species distribution and abundance models onto projection covariates for each date given.
}
\details{
Function projects a model object or list of model objects onto projection covariate data frames for each projection date given.

Projection covariate data frames must be saved “.csv” files in the drive.folder or save.drive.folder given, and must be unique in containing the relevant projection date in “YYYY-MM-DD” format. For instance, two or more “.csv” files saved with the given Google Drive folder or local directory that contain the projection date will result in function error.  Additionally, column names of projection covariate data frames must match the explanatory variable names that fitted models are trained on.

When multiple models are provided, the function projects each model onto the projection covariates and takes the average value across all model projections. However, if users have specified sdm.weight or sam.weight then the weighted average of model projections is returned. For example, this could be used to down weight projections of poorly performing models in an ensemble.

The "'proportional'" projection.method projects sdm.mod model objects onto projection covariates for each date, exporting rasters for projected distribution suitability, a continuous measure between 0 (least suitable) and 1 (most suitable).

The "'binary'" projection.method projects sdm.mod onto projection covariates for each date, exporting rasters for projected binary presence (1) or absence (0), derived from "'proportional'" projected distribution suitability using user-specified threshold (sdm.thresh) or default threshold of 0.5.

The "'abundance'" projection.method projects sam.mod onto projections covariates for each date, exporting rasters for projected abundance in the abundance units that sam.mod were fitted onto.

The "stacked" projection.method follows the "'binary'" projection.method and then projects abundance onto only binary presence (1) cells using "abundance" projection.method.

One or both of save.drive.folder and save.directory are required to specify where projection rasters are to be saved.

If one of drive.folder or save.drive.folder are used then user.email is required to access the appropriate Google Drive user account. This requires users to have installed R package "googledrive" and initialised Google Drive with valid log-in credentials. Please follow instructions on https://googledrive.tidyverse.org/.
}
