% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamic_proj_dates.R
\name{dynamic_proj_dates}
\alias{dynamic_proj_dates}
\title{Generate vector of dates for dynamic projections}
\usage{
dynamic_proj_dates(startdate, enddate, interval.level, interval)
}
\arguments{
\item{interval.level}{a character string, the time-step of intervals. One of '"day"', '"month"' or '"year"': can be abbreviated.}

\item{interval}{a numeric value, the length of interval in interval.level units to generate between the start and end date.}

\item{start.date}{a character string, the start date in format YYYY-MM-DD.}

\item{end.date}{a character string, the end date in format YYYY-MM-DD.}
}
\value{
Vector of dates between start date and end date split at regular intervals.
}
\description{
Creates a vector of dates at regular intervals between two given dates.
}
\details{
Function returns a vector of dates given start.date and end.date at given interval size.
}
