% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_xy_min_max.R
\name{extract_xy_min_max}
\alias{extract_xy_min_max}
\title{extract_xy_min_max Extracts xmin,xmax,ymin,ymax from spatial extent object}
\usage{
extract_xy_min_max(x)
}
\arguments{
\item{x}{spatial extent object. Object of class "Extent", "raster" or "polygon" or numeric vector listing xmin, xmax, ymin and ymax in order.}
}
\description{
extract_xy_min_max Extracts xmin,xmax,ymin,ymax from spatial extent object
}
