% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sample_biome_data}
\alias{sample_biome_data}
\title{Biome classification raster for southern Africa.}
\format{
A RasterLayer
\describe{
\item{class}{RasterLayer}
\item{dimensions}{4839, 3243, 15692877  (nrow, ncol, ncell)}
\item{resolution}{0.008983153, 0.008983153  (x, y)}
\item{extent}{11.71845, 40.85081, -47.89832, -4.428839  (xmin, xmax, ymin, ymax)}
\item{crs}{"+proj=longlat +datum=WGS84 +no_defs"}
\item{source}{biome_type_2001-01-01.tif }
\item{names}{biome_type_2001-01-01.tif }
\item{values}{Numerical categorical value ranging from 0, 255  (min, max).}
}
}
\usage{
sample_biome_data
}
\description{
Categorical RasterLayer cropped to southern Africa at 0.08 degree resolution from OpenLandMap
Potential Distribution of Biomes (Hengl et al., 2018).
}
\references{
Hengl T, Walsh MG, Sanderman J, Wheeler I, Harrison SP, Prentice IC. (2018) Global Mapping of
Potential Natural Vegetation: An Assessment of Machine Learning Algorithms for Estimating Land
Potential. PeerJ Preprints. <10.7287/peerj.preprints.26811v1>
}
\keyword{datasets}
