% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sample_events_data}
\alias{sample_events_data}
\title{Sample e-Bird sampling event records}
\format{
A data frame with 1000 rows and 5 variables:
\describe{
\item{day}{avian e-Bird sampling event day.}
\item{month}{avian e-Bird sampling event month.}
\item{year}{avian e-Bird sampling event year.}
\item{y}{avian e-Bird sampling event latitude.}
\item{x}{avian e-Bird sampling event longitude.}
}
}
\usage{
sample_events_data
}
\description{
A dataset containing a sample of e-Bird sampling events for all bird species across southern
Africa between 2000-2020 (Fink et al., 2021, GBIF, 2021).
The variables are as follows:
}
\references{
Fink, D., T. Auer, A. Johnston, M. Strimas-Mackey, O. Robinson, S. Ligocki, W. Hochachka, L.
Jaromczyk, C. Wood, I. Davies, M. Iliff, L. Seitz. 2021. eBird Status and Trends, Data Version:
2020; Released: 2021. Cornell Lab of Ornithology, Ithaca, New York.
\url{https://doi.org/10.2173/ebirdst.2020}

GBIF.org (12 July 2021) GBIF Occurrence Download  \url{https://doi.org/10.15468/dl.ppcu6q}
}
\keyword{datasets}
