% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sample_extent_data}
\alias{sample_extent_data}
\title{MULTIPOLYGON object for the extent of southern Africa}
\format{
A simple feature collection with 10 features and 1 field.
\describe{
\item{geometry}{MULTIPOLYGON object co-ordinates for country boundaries.}
\item{name}{name of country the polygon represents.}
}
}
\usage{
sample_extent_data
}
\description{
A MULTIPOLYGON (package "sf") object containing polygons for each country within southern Africa.
The variables are as follows:
}
\keyword{datasets}
