% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sample_filt_data}
\alias{sample_filt_data}
\title{Sample of filtered species occurrence records}
\format{
A data frame with 330 rows and 12 variables:
\describe{
\item{x}{species occurrence record x}
\item{y}{species occurrence record y}
\item{year}{species occurrence record year.}
\item{month}{species occurrence record month.}
\item{day}{species occurrence record day.}
\item{decimalLatitude}{species occurrence record latitude.}
\item{decimalLongitude}{species occurrence record longitude.}
\item{occurrenceStatus}{species presence or absence character.}
\item{source}{source of occurrence or pseudo-absence data point.}
\item{species}{name of species occurrence records belong to name}
\item{SAMP_EFFORT}{total number of avian e-Bird sampling events within spatiotemporal buffer of
occurrence record location and dates.}
\item{REL_SAMP_EFFORT}{proportion of total number of avian e-Bird sampling events within
spatiotemporal buffer of occurrence record location and dates relative to other records}
}
}
\usage{
sample_filt_data
}
\description{
A dataset containing a sample of the bird species, the red-billed quelea (Quelea quelea),
distribution records (GBIF 2021 & GBIF 2022) that have been filtered to special extent of
southern Africa and quality checked using \code{dynamicSDM} functions.
The variables are as follows:
}
\references{
GBIF.org (12 July 2021) GBIF Occurrence Download \url{https://doi.org/10.15468/dl.ppcu6q}

GBIF.org (25 July 2022) GBIF Occurrence Download \url{https://doi.org/10.15468/dl.k2kftv}
}
\keyword{datasets}
