% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sample_model_data}
\alias{sample_model_data}
\title{Sample species occurrence records with associated dynamic explanatory variables}
\format{
A data frame with 10000 rows and 14 variables:
\describe{
\item{x}{species occurrence record longitude.}
\item{y}{species occurrence record latitude.}
\item{year}{species occurrence record year.}
\item{month}{species occurrence record month.}
\item{day}{species occurrence record day.}
\item{presence.absence}{binary species presence or absence at record location and date.}
\item{individualCount}{number of individuals of the species present at record location and date.}
\item{sampling_weights}{total number of avian e-Bird sampling events within spatiotemporal buffer of occurrence record location and dates.}
\item{Temperaturemean}{mean MODIS Terra Land Surface Temperature at record co-ordinate across 52-weeks prior to record date (kelvin).}
\item{TemperatureEightmean}{mean MODIS Terra Land Surface Temperature at record co-ordinate across 8-weeks prior to record date (kelvin).}
\item{Precipitationsum}{sum Climate Hazards Group InfraRed Precipitation With Station Data (CHIRPS Daily) total daily precipitation at record co-ordinate across 52-weeks prior to record date (mm).}
\item{Precipitation8Wsum}{sum Climate Hazards Group InfraRed Precipitation With Station Data (CHIRPS Daily) total daily precipitation at record co-ordinate across 52-weeks prior to record date (mm).}
\item{watercount}{total number of MODIS Land Cover Type Yearly 500m "	Water Bodies: at least 60\% of area is covered by permanent water bodies" cells in surrounding area of record co-ordinate in record year.}
\item{blockno}{Block group assinged for model fitting by jacknife approach to account for spatial and temporal autocorrelation}
}
}
\description{
A dataset containing a sample of the bird species, the red-billed quelea (Quelea quelea), distribution and abundance records from between 2002-2017 including generated pseudo-absence records and associated extracted dynamic explanatory variables.
The variables are as follows:
}
\references{
Friedl, M., Sulla-Menashe, D. (2019). MCD12Q1 MODIS/Terra+Aqua Land Cover Type Yearly L3 Global 500m SIN Grid V006 \link{Data set}. NASA EOSDIS Land Processes DAAC. Accessed 2022-11-24 from https://doi.org/10.5067/MODIS/MCD12Q1.006

Funk, Chris, Pete Peterson, Martin Landsfeld, Diego Pedreros, James Verdin, Shraddhanand Shukla, Gregory Husak, James Rowland, Laura Harrison, Andrew Hoell & Joel Michaelsen. "The climate hazards infrared precipitation with stations-a new environmental record for monitoring extremes". Scientific Data 2, 150066. doi:10.1038/sdata.2015.66 2015.

GBIF.org (12 July 2021) GBIF Occurrence Download https://doi.org/10.15468/dl.ppcu6q

GBIF.org (25 July 2022) GBIF Occurrence Download https://doi.org/10.15468/dl.k2kftv

Wan, Z., Hook, S., Hulley, G. (2015). MOD11A1 MODIS/Terra Land Surface Temperature/Emissivity Daily L3 Global 1km SIN Grid V006 \link{Data set}. NASA EOSDIS Land Processes DAAC. Accessed 2022-11-24 from https://doi.org/10.5067/MODIS/MOD11A1.006
}
