% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sample_occ_abs_data}
\alias{sample_occ_abs_data}
\title{Sample species occurrence records with pseudo-absence records}
\format{
A data frame with 50 rows and 7 variables:
\describe{
\item{x}{species occurrence record longitude.}
\item{y}{species occurrence record latitude.}
\item{year}{species occurrence record year.}
\item{month}{species occurrence record month.}
\item{day}{species occurrence record day.}
\item{presence.absence}{binary species presence or absence at record location and date.}
\item{individualCount}{number of individuals of the species present at record location and date.}
}
}
\usage{
sample_occ_abs_data
}
\description{
A dataset containing a sample of the bird species, the red-billed quelea (Quelea quelea), distribution and abundance records between 2002-2017 with generated spatiotemporal pseudo-absence records.
The variables are as follows:
}
\references{
GBIF.org (12 July 2021) GBIF Occurrence Download https://doi.org/10.15468/dl.ppcu6q

GBIF.org (25 July 2022) GBIF Occurrence Download https://doi.org/10.15468/dl.k2kftv
}
\keyword{datasets}
