% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sample_proj_rast}
\alias{sample_proj_rast}
\title{Distribution suitability projections for red-billed quelea in southern Africa}
\format{
A RasterBrick
\describe{
\item{class}{RasterBrick}
\item{dimensions}{31, 30, 930, 4  (nrow, ncol, ncell, nlayers)}
\item{resolution}{1, 1  (x, y)}
\item{extent}{11.6901, 41.6901, -35.43802, -4.438023  (xmin, xmax, ymin, ymax)}
\item{crs}{"+proj=longlat +datum=WGS84 +no_defs"}
\item{source}{memory} \item{names}{ X2018.01.01_proportional, X2018.04.01_proportional,
X2018.07.01_proportional, X2018.10.01_proportional }
\item{min values}{Numerical minimum distribution suitability projected in each layer.}
\item{max values}{Numerical maximum distribution suitability projected in each layer.}
}
}
\usage{
sample_proj_rast
}
\description{
RasterBrick of distribution suitability projections for red-billed quelea across southern at one
degree resolution for the dates: "2018-01-01", "2018-04-01", "2018-07-01" and "2018-10-01". A
value of 1 represents high suitability and a value of 0 low suitability.
}
\keyword{datasets}
