% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GEE_reducer_analogous.R
\name{skip_if_no_GEE_credentials}
\alias{skip_if_no_GEE_credentials}
\title{skip_if_no_GEE_credentials Skips test_that test if no Google Earth Engine credentials set-up}
\usage{
skip_if_no_GEE_credentials()
}
\description{
skip_if_no_GEE_credentials Skips test_that test if no Google Earth Engine credentials set-up
}
