% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatiotemp_check.R
\name{spatiotemp_check}
\alias{spatiotemp_check}
\title{Check species occurrence record formatting, completeness and validity.}
\usage{
spatiotemp_check(
  occ.data,
  na.handle = "exclude",
  duplicate.handle = "exclude",
  coord.handle = "exclude",
  date.handle = "exclude"
)
}
\arguments{
\item{occ.data}{a data frame, with columns for occurrence record co-ordinates and dates with column names as follows; record longitude as "x", latitude as "y", year as "year", month as "month", and day as "day".}

\item{na.handle}{a character string, method for handling missing data (NA values) in record co-ordinates and dates. One of '"exclude"' or '"ignore"': can be abbreviated. Default; exclude.}

\item{duplicate.handle}{a character string, method for handling duplicate record co-ordinates or dates. One of '"exclude"' or '"ignore"': can be abbreviated. Default; exclude.}

\item{coord.handle}{a character string, method for handling invalid co-ordinates in record data. One of '"exclude"' or '"ignore"': can be abbreviated. Default; exclude.}

\item{date.handle}{a character string, method for handling invalid dates in record data. One of '"exclude"' or '"ignore"': can be abbreviated. Default; exclude.}
}
\value{
By default, returns occurrence record data frame, filtered to exclude records containing missing, duplicate or invalid data in record co-ordinates and dates.
}
\description{
Checks the occurrence record data frame contains the column names and classes required for dynamicSDM functions. Option to exclude records containing missing, duplicate or invalid co-ordinates or dates.
}
\details{
Record dates and co-ordinates are checked for validity using the following rules. Dates must be real dates that could exist. For example, 50th February 2000 is not a valid date. Co-ordinates must have longitude (x) values between -180 and 180, and latitude (y) values between -90 and 90 to be considered valid.
}
\examples{
data("sample_occ_data")
spatiotemp_check(sample_occ_data)
}
