% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatiotemp_extent.R
\name{spatiotemp_extent}
\alias{spatiotemp_extent}
\title{Filter species occurrence records by a given spatial and temporal extent.}
\usage{
spatiotemp_extent(occ.data, temporal.ext = NULL, spatial.ext = NULL)
}
\arguments{
\item{occ.data}{a data frame, with columns for occurrence record co-ordinates and dates with column names as follows; record longitude as "x", latitude as "y", year as "year", month as "month", and day as "day".}

\item{temporal.ext}{optional; a character vector, two dates in format YYYY-MM-DD. First date represents start of temporal extent and second date represents end of temporal extent for inclusion.}

\item{spatial.ext}{optional; object of class "Extent", "raster" or "polygon" or numeric vector listing xmin, xmax, ymin and ymax in order, the spatial extent to filter by.}
}
\value{
Returns data frame of occurrence records filtered to the spatial and temporal extent given.
}
\description{
Function excludes species occurrence records with co-ordinates outside given spatial extent and dates outside given temporal extent.
}
\details{
If spatial.ext is provided, the function checks whether species occurrence record co-ordinates are within the given spatial extent of the study (spatial.ext) and excludes any outside of this extent.

If temporal.ext is provided, the function checks whether species occurrence record dates are within the given temporal extent of the study (temporal.ext) and excludes any outside of this extent.
}
\examples{
x<-c(27.79125, 28.54125, 25.54125, 30.04125, 29.95792)
y<-c(-26.79125, -26.37458, -26.70792, -29.37458, -28.45792)
month<-c(1, 2, 3, 2, 4)
day<-c(27, 25, 16, 25, 26)
year<-c(2014, 2016, 2011, 2011, 2015)
occ.data<-data.frame(cbind(x,y,year,month,day))
spatiotemp_extent(occ.data,
                 temporal.ext=c("2012-01-01","2017-01-01"),
                spatial.ext =c(28,31,-30,-26))
}
