% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamicSDM-package.R
\docType{package}
\name{dynamicSDM-package}
\alias{dynamicSDM}
\alias{dynamicSDM-package}
\title{dynamicSDM: Dynamic species distribution and abundance modelling}
\description{
This package enables the generation of species distribution and abundance models that are dynamic through space and time. dynamicSDM presents highly-flexible functions for incorporating both spatial and temporal aspects at key species distribution modelling stages, including when cleaning and filtering species occurrence data, assessing and correcting biases and autocorrelation, generating pseudo-absence records, extracting dynamic explanatory variables and projecting species distribution and abundance patterns at high spatiotemporal resolution. Functions take advantage of Google Earth Engine and Google Drive to minimise computing power and storage demands.
}
\author{
\strong{Maintainer}: Rachel Dobson \email{eerdord@gmail.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
