% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_buffered_coords.R
\name{extract_buffered_coords}
\alias{extract_buffered_coords}
\title{Extract spatially buffered and temporally dynamic explanatory variable data for occurrence records.}
\usage{
extract_buffered_coords(
  occ.data,
  datasetname,
  bandname,
  spatial.res.metres,
  GEE.math.fun,
  moving.window.matrix,
  extraction.drive.folder,
  user.email,
  save.method,
  varname = NULL,
  temporal.res = NULL,
  temporal.level = NULL,
  temporal.direction = NULL,
  categories = NULL,
  save.directory = NULL,
  agg.factor = NULL
)
}
\arguments{
\item{occ.data}{a data frame, with columns for occurrence record co-ordinates and dates with column names as follows; record longitude as "x", latitude as "y", year as "year", month as "month", and day as "day".}

\item{datasetname}{a character string, the Google Earth Engine dataset to extract data from.}

\item{bandname}{a character string, the Google Earth Engine dataset bandname to extract data for.}

\item{spatial.res.metres}{a numeric value, the spatial resolution in metres for data extraction.}

\item{GEE.math.fun}{a character string, the mathematical function to compute across the specified spatial matrix and period for each record.}

\item{moving.window.matrix}{a matrix of weights with an odd number of sides, representing the spatial neighbourhood of cells (“moving window”) to calculate GEE.math.fun across from record co-ordinate. See details for more information.}

\item{extraction.drive.folder}{a character string, Google Drive folder to temporarily save rasters to using Google Earth Engine before spatial buffering in R.}

\item{user.email}{a character string, user email for initialising Google Drive.}

\item{save.method}{a character string, the method used to save extracted variable data. One of '"split"' or '"combined"': can be abbreviated. See details.}

\item{varname}{optional; a character string, a unique name for the explanatory variable. Default varname is “bandname_temporal.res_temporal.direction_ GEE.math.fun_buffered".}

\item{temporal.res}{optional; a numeric value, the temporal resolution in days to extract data and calculate GEE.math.fun across from occurrence record date.}

\item{temporal.level}{a character string, the temporal resolution of the explanatory variable data. One of '"day"', '"month"' or '"year"': can be abbreviated. Default; day.}

\item{temporal.direction}{optional; a character string, the temporal direction for extracting data across relative to the record date. One of '"prior"' or '"post"': can be abbreviated.}

\item{categories}{optional; a character string, the categories to use in calculation if data are categorical. See details for more information.}

\item{save.directory}{a character string, path to a local directory to save extracted variable data to.}

\item{agg.factor}{optional; a numeric value, the factor to aggregate data by before spatial buffering.}
}
\value{
Returns details of successful explanatory variable extractions.
}
\description{
For each species occurrence record co-ordinate and date, spatially buffered and temporally dynamic explanatory data are extracted using Google Earth Engine.
}
\details{
For each individual species occurrence record co-ordinate and date, this function extracts data for a given band within a Google Earth Engine dataset across a user-specified spatial buffer and temporal period and calculates a mathematical function on such data.

If temporal.res and temporal.direction are not given, the function extracts explanatory variable data (in RasterLayer format) for all of the cells surrounding and including the cell containing the occurrence record co-ordinates. If temporal.res and temporal.direction is given, the function extracts explanatory variable data (in RasterLayer format) for which GEE.math.fun has been first calculated over this period in relation to the occurrence record date.

Using the focal function in raster R package (Hijmans et al., 2015), GEE.math.fun is calculated across the spatial buffer area from the record co-ordinate. The spatial buffer area used is specified by argument moving.window matrix, which dictates the neighbourhood of cells surrounding the cell containing the occurrence record to include in the calculation.

When explanatory variable data are categorical (e.g. land cover classes), argument categories can be used to specify the categories of importance to the calculation. The category or categories given will be converted in a binary representation, with “1” for those listed, and “0” for all others in the dataset. Ensure that the GEE.math.fun given is appropriate for such data. For example, this function could return the sum of suitable land cover classified cells in the “moving window” from species occurrence record co-ordinates.

extract_buffered_coords requires users to have installed R package "rgee" (Aybar et al., 2020) and initialised Google Earth Engine with valid log-in credentials. Please follow instructions on the following website https://cran.r-project.org/web/packages/rgee/vignettes/rgee01.html. datasetname must be in the accepted Google Earth Engine Data catalogue layout (e.g. “MODIS/006/MCD12Q1” or “UCSB-CHG/CHIRPS/DAILY”) and bandname as specified in the dataset (e.g. “LC_Type5”, “precipitation”). For datasets and band names, see https://developers.google.com/earth-engine/datasets.

extract_buffered_coords also requires users to have installed the R package "googledrive" (D'Agostino McGowan and Bryan, 2022) and initialised Google Drive with valid log-in credentials, which must be provided under argument user.email. Please follow instructions on https://googledrive.tidyverse.org/ for initialising the googledrive package.

GEE.math.fun specifies the mathematical function to be calculated over the spatial buffered area and temporal period. Options are limited to Google Earth Engine ImageCollection Reducer functions (https://developers.google.com/earth-engine/apidocs/) for which an analogous R function is available. This includes: "allNonZero","anyNonZero", "count", "first","firstNonNull", "last", "lastNonNull", "max","mean", "median","min", "mode","product", "sampleStdDev", "sampleVariance", "stdDev", "sum" and "variance".

temporal.level states the temporal resolution of the explanatory variable data and improves the speed of extract_buffered_coords extraction. For example, if the explanatory data represents an annual variable, then all record co-ordinates from the same year can be extracted from the same raster. However, if the explanatory data represents a daily variable, then only records from the exact same day can be extracted from the same raster. For the former, temporal.level argument should be ‘year’ and for the latter, temporal.level should be ‘day’.

For save.method '"combined"', the function with save “.csv” files containing all occurrence records and associated values for the explanatory variable. If save.method '"split"' is chosen, the function will save individual “.csv” files for all of the records with each unique period of the given temporal.level (e.g. each year, each year and month combination or each unique date). '"split"' method is provided to protect users if internet connection is lost when extracting data for large occurrence datasets.
}
\references{
Aybar, C., Wu, Q., Bautista, L., Yali, R. and Barja, A., 2020. rgee: An R package for interacting with Google Earth Engine. Journal of Open Source Software, 5(51), p.2272.

D'Agostino McGowan L., and Bryan J., 2022. googledrive: An Interface to Google Drive. https://googledrive.tidyverse.org, https://github.com/tidyverse/googledrive.

Hijmans, R. J., Van Etten, J., Cheng, J., Mattiuzzi, M., Sumner, M., Greenberg, J. A., Lamigueiro, O. P., Bevan, A., Racine, E. B. & Shortridge, A. 2015. Package ‘raster’. R package, 734.
}
