% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_coords_combine.R
\name{extract_coords_combine}
\alias{extract_coords_combine}
\title{Combine extracted explanatory variable data for occurrence records into single data frame.}
\usage{
extract_coords_combine(varnames, local.directory)
}
\arguments{
\item{varnames}{a character string, the unique names for each explanatory variable.}

\item{local.directory}{a character string or vector, the path to local directory or directories to read extracted explanatory data frames from.}
}
\value{
Returns a data frame containing all occurrence records with associated explanatory variable data.
}
\description{
Combines the split output files from functions extract_dynamic_coords and extract_buffered_coords into single data frame containing all occurrence records and explanatory variables.
}
\details{
When extract_dynamic_coords and extract_buffered_coords have been used to extract dynamic explanatory variables for occurrence records, output for individual records and each variable will be split into separate “.csv” files.

This function reads in these files and combined data into a single data frame containing each occurrence records and associated explanatory data from each variable.

To prevent error, the “.csv” files must be uniquely named within the folder and include an exact character match for the varnames provided. All “.csv” files matching the varnames should have the same number and names of columns, which the default output of extract_dynamic_coords and extract_buffered_coords will if given the same occ.data data frame.
}
